Updated to MAME 0.116u4 core; many changes with timers and bioses; watch out for regressions
git-svn-id: svn://messdev.no-ip.org/mess@683 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mame/machine/6526cia.h
@@ -23,8 +23,8 @@ struct _cia6526_interface
 {
 	cia_type_t type;
 	void (*irq_func)(int state);
-	double clock;
-	double tod_clock;
+	int clock;
+	int tod_clock;
 
 	struct
 	{
